renamed some classes name
