+++ docs/dev/stream/side_output.md
@@ -60,7 +60,7 @@ Emitting data to a side output is possible from the following functions:
 - [ProcessFunction]({{ site.baseurl }}/dev/stream/operators/process_function.html)
 - [KeyedProcessFunction]({{ site.baseurl }}/dev/stream/operators/process_function.html#the-keyedprocessfunction)
 - CoProcessFunction
-- CoKeyedProcessFunction
+- KeyedCoProcessFunction
 - [ProcessWindowFunction]({{ site.baseurl }}/dev/stream/operators/windows.html#processwindowfunction)
 - ProcessAllWindowFunction
