+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/datastream/ConnectedStreams.java
@@ -26,15 +26,15 @@ import org.apache.flink.api.java.functions.KeySelector;
 import org.apache.flink.api.java.typeutils.TypeExtractor;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.streaming.api.functions.co.CoFlatMapFunction;
-import org.apache.flink.streaming.api.functions.co.CoKeyedProcessFunction;
 import org.apache.flink.streaming.api.functions.co.CoMapFunction;
 import org.apache.flink.streaming.api.functions.co.CoProcessFunction;
+import org.apache.flink.streaming.api.functions.co.KeyedCoProcessFunction;
 import org.apache.flink.streaming.api.operators.TwoInputStreamOperator;
 import org.apache.flink.streaming.api.operators.co.CoProcessOperator;
 import org.apache.flink.streaming.api.operators.co.CoStreamFlatMap;
 import org.apache.flink.streaming.api.operators.co.CoStreamMap;
 import org.apache.flink.streaming.api.operators.co.KeyedCoProcessOperator;
-import org.apache.flink.streaming.api.operators.co.PureKeyedCoProcessOperator;
+import org.apache.flink.streaming.api.operators.co.LegacyKeyedCoProcessOperator;
 import org.apache.flink.streaming.api.transformations.TwoInputTransformation;
 
 import static java.util.Objects.requireNonNull;
@@ -333,7 +333,7 @@ public class ConnectedStreams<IN1, IN2> {
 		TwoInputStreamOperator<IN1, IN2, R> operator;
 
 		if ((inputStream1 instanceof KeyedStream) && (inputStream2 instanceof KeyedStream)) {
-			operator = new KeyedCoProcessOperator<>(inputStream1.clean(coProcessFunction));
+			operator = new LegacyKeyedCoProcessOperator<>(inputStream1.clean(coProcessFunction));
 		} else {
 			operator = new CoProcessOperator<>(inputStream1.clean(coProcessFunction));
 		}
@@ -342,7 +342,7 @@ public class ConnectedStreams<IN1, IN2> {
 	}
 
 	/**
-	 * Applies the given {@link CoKeyedProcessFunction} on the connected input keyed streams,
+	 * Applies the given {@link KeyedCoProcessFunction} on the connected input keyed streams,
 	 * thereby creating a transformed output stream.
 	 *
 	 * <p>The function will be called for every element in the input keyed streams and can produce zero or
@@ -350,7 +350,7 @@ public class ConnectedStreams<IN1, IN2> {
 	 * function can also query the time and set timers. When reacting to the firing of set timers
 	 * the function can directly emit elements and/or register yet more timers.
 	 *
-	 * @param coKeyedProcessFunction The {@link CoKeyedProcessFunction} that is called for each element
+	 * @param keyedCoProcessFunction The {@link KeyedCoProcessFunction} that is called for each element
 	 *                      in the stream.
 	 *
 	 * @param <R> The type of elements emitted by the {@code CoProcessFunction}.
@@ -359,11 +359,11 @@ public class ConnectedStreams<IN1, IN2> {
 	 */
 	@PublicEvolving
 	public <K, R> SingleOutputStreamOperator<R> process(
-		CoKeyedProcessFunction<K, IN1, IN2, R> coKeyedProcessFunction) {
+		KeyedCoProcessFunction<K, IN1, IN2, R> keyedCoProcessFunction) {
 
 		TypeInformation<R> outTypeInfo = TypeExtractor.getBinaryOperatorReturnType(
-			coKeyedProcessFunction,
-			CoKeyedProcessFunction.class,
+			keyedCoProcessFunction,
+			KeyedCoProcessFunction.class,
 			0,
 			1,
 			2,
@@ -373,11 +373,11 @@ public class ConnectedStreams<IN1, IN2> {
 			Utils.getCallLocationName(),
 			true);
 
-		return process(coKeyedProcessFunction, outTypeInfo);
+		return process(keyedCoProcessFunction, outTypeInfo);
 	}
 
 	/**
-	 * Applies the given {@link CoKeyedProcessFunction} on the connected input streams,
+	 * Applies the given {@link KeyedCoProcessFunction} on the connected input streams,
 	 * thereby creating a transformed output stream.
 	 *
 	 * <p>The function will be called for every element in the input streams and can produce zero
@@ -385,7 +385,7 @@ public class ConnectedStreams<IN1, IN2> {
 	 * this function can also query the time and set timers. When reacting to the firing of set
 	 * timers the function can directly emit elements and/or register yet more timers.
 	 *
-	 * @param coKeyedProcessFunction The {@link CoKeyedProcessFunction} that is called for each element
+	 * @param keyedCoProcessFunction The {@link KeyedCoProcessFunction} that is called for each element
 	 *                      in the stream.
 	 *
 	 * @param <R> The type of elements emitted by the {@code CoProcessFunction}.
@@ -394,15 +394,15 @@ public class ConnectedStreams<IN1, IN2> {
 	 */
 	@Internal
 	public <K, R> SingleOutputStreamOperator<R> process(
-		CoKeyedProcessFunction<K, IN1, IN2, R> coKeyedProcessFunction,
+		KeyedCoProcessFunction<K, IN1, IN2, R> keyedCoProcessFunction,
 		TypeInformation<R> outputType) {
 
 		TwoInputStreamOperator<IN1, IN2, R> operator;
 
 		if ((inputStream1 instanceof KeyedStream) && (inputStream2 instanceof KeyedStream)) {
-			operator = new PureKeyedCoProcessOperator<>(inputStream1.clean(coKeyedProcessFunction));
+			operator = new KeyedCoProcessOperator<>(inputStream1.clean(keyedCoProcessFunction));
 		} else {
-			throw new UnsupportedOperationException("CoKeyedProcessFunction can only be used " +
+			throw new UnsupportedOperationException("KeyedCoProcessFunction can only be used " +
 				"when both input streams are of type KeyedStream.");
 		}
